dp3/ingestion/requirements.txt
requests==2.32.3
boto3==1.37.38
Pillow==11.2.1
